traffic-analyzer/traffic-analyzer-kimi/requirements.txt
# Traffic Analyzer - 网络流量分析工具依赖

# 数据分析
pandas>=1.5.0
numpy>=1.24.0

# 科学计算
scipy>=1.9.0

# 配置文件解析
PyYAML>=6.0

# 进度条和CLI美化
tqdm>=4.65.0
colorama>=0.4.6

# 可选：可视化（扩展功能）
# matplotlib>=3.6.0
# seaborn>=0.12.0
